Changed caching strategy to make indexers more efficient.
I changed CheckMemberCache to return a bool indicating whether the
member was in the cache so that I could allow a null member to be
stored in the cache. Doing so allowed me to indicate in the cache that
a search had already been made for an item but was not found, thereby
eliminating the need to search again. Hence, the fallback to the
indexer is much faster because the cache indicates whether the methods
and extensions methods were already searched for a particular index,
eliminating a host of unnecessary reflection after the first access.