Revised caching algorithm for improving indexer performance.
Changed the caching algorithm to cache the names and MethodInfo’s of
all extension methods, and then to mark the fact that extension methods
have been searched in the cache. This allows the extension method
search to be skipped immediately if the method name does not match an
extension method, without doing reflection. Since the indexer checks
necessarily follow the extension method checks, this speeds up indexer
performance dramatically.