Fixed issue where variadic static method arguments wouldn't get read from cache properly
* The count of cached parameters was done wrong (I assume), so the parameter array was empty when calling a static method more than once with a variadic parameter array for some reasons.

The best way to spot the issues immediately is to try to wrap "String.Format(string, params object[])" in a static class method "Format(string, params object[])". And then call it twice in a row with any valid paramters. The second time, the parameter array will be completely empty even if data was passed to the function inside the lua script.