Wrap C functions to throw errors in Lua functions
This change makes all C functions not throw their Lua errors directly,
and instead makes them return it for a Lua wrapper to pick up and raise.
By doing this, the issue of longjmp translation is bypassed/fixed on
non-Windows platforms, as the errors are thrown and caught inside native
Lua runtime code.

Fixes #344